fix some more issues
