+++ core/spec/models/spree/shipping_method_spec.rb
@@ -14,12 +14,12 @@ describe Spree::ShippingMethod, type: :model do
       allow(Spree::ShippingMethod).to receive_message_chain(:spree_calculators, :shipping_methods).and_return([
                                                                                                                 Spree::Calculator::Shipping::FlatPercent,
                                                                                                                 Spree::Calculator::Shipping::PriceSack,
-                                                                                                                Spree::Calculator::Tax::DefaultTax,
+                                                                                                                Spree::Calculator::DefaultTax,
                                                                                                                 DummyShippingCalculator # included as regression test for https://github.com/spree/spree/issues/3109
                                                                                                               ])
 
       expect(Spree::ShippingMethod.calculators).to eq([Spree::Calculator::Shipping::FlatPercent, Spree::Calculator::Shipping::PriceSack, DummyShippingCalculator])
-      expect(Spree::ShippingMethod.calculators).not_to eq([Spree::Calculator::Tax::DefaultTax])
+      expect(Spree::ShippingMethod.calculators).not_to eq([Spree::Calculator::DefaultTax])
     end
   end
