+++ sample/db/samples/tax_rates.rb
@@ -7,5 +7,5 @@ Spree::TaxRate.where(
   amount: 0.05,
   tax_category: clothing
 ).first_or_create! do |tax_rate|
-  tax_rate.calculator = Spree::Calculator::Tax::DefaultTax.create!
+  tax_rate.calculator = Spree::Calculator::DefaultTax.create!
 end
